+++ capsule/src/main/java/ca/allanwang/capsule/library/swiperecyclerview/managers/SGridLayoutManager.java
@@ -1,10 +1,13 @@
 package ca.allanwang.capsule.library.swiperecyclerview.managers;
 
 import android.content.Context;
+import android.graphics.PointF;
+import android.support.annotation.Nullable;
 import android.support.v7.widget.GridLayoutManager;
 import android.support.v7.widget.LinearSmoothScroller;
 import android.support.v7.widget.RecyclerView;
 import android.util.AttributeSet;
+import android.util.DisplayMetrics;
 
 import ca.allanwang.capsule.library.swiperecyclerview.interfaces.ILayoutManager;
 
@@ -14,7 +17,7 @@ import ca.allanwang.capsule.library.swiperecyclerview.interfaces.ILayoutManager;
 
 public class SGridLayoutManager extends GridLayoutManager implements ILayoutManager {
     private boolean isScrollEnabled = true;
-    private ScrollTime scrollTime; //One time switch
+    private ScrollSpeed scrollSpeed; //One time switch
 
     public SGridLayoutManager(Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes) {
         super(context, attrs, defStyleAttr, defStyleRes);
@@ -34,22 +37,29 @@ public class SGridLayoutManager extends GridLayoutManager implements ILayoutMana
     }
 
     @Override
-    public void setSmoothScrollDuration(ScrollTime scrollTime) {
-        this.scrollTime = scrollTime;
+    public void setSmoothScrollDuration(ScrollSpeed scrollSpeed) {
+        this.scrollSpeed = scrollSpeed;
     }
 
     @Override
     public void smoothScrollToPosition(RecyclerView recyclerView, RecyclerView.State state, int position) {
         LinearSmoothScroller linearSmoothScroller =
                 new LinearSmoothScroller(recyclerView.getContext()) {
+
                     @Override
-                    protected int calculateTimeForScrolling(int dx) {
-                        if (scrollTime != null) {
-                            int duration = scrollTime.calculateTimeForScrolling(dx, super.calculateTimeForScrolling(dx));
-                            scrollTime = null; //Reset
-                            return duration;
+                    protected float calculateSpeedPerPixel(DisplayMetrics displayMetrics) {
+                        if (scrollSpeed != null) {
+                            float speed = scrollSpeed.calculateSpeedPerPixel(super.calculateSpeedPerPixel(displayMetrics));
+                            scrollSpeed = null;
+                            return speed;
                         }
-                        return super.calculateTimeForScrolling(dx);
+                        return super.calculateSpeedPerPixel(displayMetrics);
+                    }
+
+                    @Nullable
+                    @Override
+                    public PointF computeScrollVectorForPosition(int targetPosition) {
+                        return super.computeScrollVectorForPosition(targetPosition);
                     }
                 };
         linearSmoothScroller.setTargetPosition(position);