+++ capsule/src/main/java/ca/allanwang/capsule/library/swiperecyclerview/managers/SLinearLayoutManager.java
@@ -1,12 +1,15 @@
 package ca.allanwang.capsule.library.swiperecyclerview.managers;
 
 import android.content.Context;
+import android.graphics.PointF;
+import android.support.annotation.Nullable;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.LinearSmoothScroller;
 import android.support.v7.widget.RecyclerView;
 import android.util.AttributeSet;
 import android.util.DisplayMetrics;
 
+import ca.allanwang.capsule.library.logging.CLog;
 import ca.allanwang.capsule.library.swiperecyclerview.interfaces.ILayoutManager;
 
 /**
@@ -15,7 +18,7 @@ import ca.allanwang.capsule.library.swiperecyclerview.interfaces.ILayoutManager;
 
 public class SLinearLayoutManager extends LinearLayoutManager implements ILayoutManager {
     private boolean isScrollEnabled = true;
-    private ScrollTime scrollTime; //One time switch
+    private ScrollSpeed scrollSpeed; //One time switch
 
     public SLinearLayoutManager(Context context) {
         super(context);
@@ -35,22 +38,29 @@ public class SLinearLayoutManager extends LinearLayoutManager implements ILayout
     }
 
     @Override
-    public void setSmoothScrollDuration(ScrollTime scrollTime) {
-        this.scrollTime = scrollTime;
+    public void setSmoothScrollDuration(ScrollSpeed scrollSpeed) {
+        this.scrollSpeed = scrollSpeed;
     }
 
     @Override
     public void smoothScrollToPosition(RecyclerView recyclerView, RecyclerView.State state, int position) {
         LinearSmoothScroller linearSmoothScroller =
                 new LinearSmoothScroller(recyclerView.getContext()) {
+
                     @Override
-                    protected int calculateTimeForScrolling(int dx) {
-                        if (scrollTime != null) {
-                            int duration = scrollTime.calculateTimeForScrolling(dx, super.calculateTimeForScrolling(dx));
-                            scrollTime = null; //Reset
-                            return duration;
+                    protected float calculateSpeedPerPixel(DisplayMetrics displayMetrics) {
+                        if (scrollSpeed != null) {
+                            float speed = scrollSpeed.calculateSpeedPerPixel(super.calculateSpeedPerPixel(displayMetrics));
+                            scrollSpeed = null;
+                            return speed;
                         }
-                        return super.calculateTimeForScrolling(dx);
+                        return super.calculateSpeedPerPixel(displayMetrics);
+                    }
+
+                    @Nullable
+                    @Override
+                    public PointF computeScrollVectorForPosition(int targetPosition) {
+                        return SLinearLayoutManager.this.computeScrollVectorForPosition(targetPosition);
                     }
                 };
         linearSmoothScroller.setTargetPosition(position);